feat: Added canvas dragging
app.js :
Created a canvas dragging functionality that allows the user to move the canvas in the desired direction if the user clicks & drags the canvas, and has clicked on the cursor button.

index.html :
Added two buttons, one for activating the painting functionality and the other for activating the click & drag functionality.